requirements.txt
Flask
flask_socketio
fyers_apiv3
SQLAlchemy
polars
gunicorn
eventlet
python-dotenv
setuptools